Corrigido erro no /vanish, adicionado /fly

+++ src/main/java/com/pwdim/tutorial/listener/ScoreBoardListener.java
@@ -7,8 +7,11 @@ import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerInteractEvent;
 import org.bukkit.event.player.PlayerJoinEvent;
+import org.bukkit.event.player.PlayerMoveEvent;
 import org.bukkit.event.player.PlayerQuitEvent;
+import org.bukkit.plugin.Plugin;
 
 import java.util.Collection;
 import java.util.HashMap;
@@ -19,6 +22,7 @@ import java.util.stream.Collectors;
 public class ScoreBoardListener implements Listener {
     private final TUTORIAL plugin;
     private final Map<UUID, FastBoard> boards = new HashMap<>();
+
 
     public ScoreBoardListener(TUTORIAL plugin) {
         this.plugin = plugin;
@@ -32,9 +36,29 @@ public class ScoreBoardListener implements Listener {
         String footer = ColorUtils.color("&b&oplay.pwdim.com");
         String online = ColorUtils.color("&fOnline: &b&o"+ plugin.getOnlinePlayers());
 
+        plugin.getBoards().put(p.getUniqueId()  , board);
+
         board.updateTitle(title);
         board.updateLine(0, ColorUtils.color(" "));
-        board.updateLine(1, ColorUtils.color("&fOlá, %s.", p.getCustomName()));
+        board.updateLine(1, ColorUtils.color("&fOlá, %s&f.", p.getCustomName()));
+        board.updateLine(2, ColorUtils.color(" "));
+        board.updateLine(3, ColorUtils.color(online));
+        board.updateLine(4, ColorUtils.color(" "));
+        board.updateLine(5, footer);
+    }
+    @EventHandler
+    public void setScoreBoard(PlayerMoveEvent e) {
+        Player p = e.getPlayer();
+        FastBoard board = new FastBoard(p);
+        String title = ColorUtils.color("&B&LPWDIM");
+        String footer = ColorUtils.color("&b&oplay.pwdim.com");
+        String online = ColorUtils.color("&fOnline: &b&o"+ plugin.getOnlinePlayers());
+
+        plugin.getBoards().put(p.getUniqueId()  , board);
+
+        board.updateTitle(title);
+        board.updateLine(0, ColorUtils.color(" "));
+        board.updateLine(1, ColorUtils.color("&fOlá, %s&f.", p.getCustomName()));
         board.updateLine(2, ColorUtils.color(" "));
         board.updateLine(3, ColorUtils.color(online));
         board.updateLine(4, ColorUtils.color(" "));
@@ -44,8 +68,12 @@ public class ScoreBoardListener implements Listener {
     @EventHandler
     public void onQuit(PlayerQuitEvent e) {
         Player p = e.getPlayer();
-        FastBoard board = new FastBoard(p);
 
-        board.delete();
+        FastBoard board = plugin.getBoards().remove(p.getUniqueId());
+        if (board != null) {
+            board.delete();
+        }
     }
+
+
 }
